🎉 Complete FacilityManager modularization success
✅ Extracted 6 components + custom hook:
- CorporateInfoForm.jsx
- FacilitySearch.jsx
- SelectedFacilitiesList.jsx
- ManualFacilityEntry.jsx
- PSAReviewAndSigning.jsx
- OrganizationTypeSelector.jsx
- useFacilityManager.js custom hook

✅ 78% code reduction (1,400 → 300 lines)
✅ All functionality preserved and tested working
✅ ACR search, facility management, auto-population all working
✅ Ready for patient dashboard reuse

🚀 Major architecture improvement complete!
